pyproject.toml
[project]
name = "objectstoragemcp"
version = "0.1.0"
description = "MCP сервер для работы с AWS S3 Object Storage: загрузка, скачивание, листинг и управление файлами"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.13.3",
    "python-dotenv>=1.2.1",
    "boto3>=1.35.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.0",
    "moto>=5.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
